R6 Class for Omicidx Database Interaction

The `Omicidx` R6 class provides an interface for interacting with the omicidx database.

Public fields

connection

A DBI connection object to the omicidx database.

.dplyr_tbls

A list to cache dplyr table objects.

Methods

Public methods

Omicidx$new()

Omicidx$sql()

Omicidx$table_names()

Omicidx$tbl()

Omicidx$clone()

Method new()

Create a new `Omicidx` object

Usage

Omicidx$new(con = NULL)

Arguments

con

An optional DBI connection object. If NULL, a new connection will be created.

Returns

A new `Omicidx` object.

Method sql()

Execute a SQL query on the omicidx database.

Usage

Omicidx$sql(query)

Arguments

query

A string containing the SQL query to execute.

Returns

A data frame containing the results of the query.

Method table_names()

List all table names in the omicidx database.

Usage

Omicidx$table_names()

Returns

A data frame containing the names of all tables in the database.

Method tbl()

Get a dplyr table object for a specified table in the omicidx database.

Usage

Omicidx$tbl(table_name)

Arguments

table_name

A string specifying the name of the table.

Details

This method retrieves a dplyr table object for the specified table name. If the table has been previously accessed, it will return the cached version. If the table_name does not exist, an error will be raised.

Returns

A dplyr table object for the specified table.

Method clone()

The objects of this class are cloneable with this method.

Usage

Omicidx$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# # Create a new Omicidx object
omicidx <- Omicidx$new()
#> Error: object 'Omicidx' not found
# # List all table names in the database
omicidx$table_names()
#> Error: object 'omicidx' not found
# Get a dplyr table object for a specific table
sra_studies_tbl <- omicidx$tbl("sra_studies")
#> Error: object 'omicidx' not found
# Perform a dplyr query
result <- sra_studies_tbl |>
  filter(study_type == "RNA-Seq") |>
  head(10) |>
  collect()
#> Error: object 'sra_studies_tbl' not found
print(result)
#> Error: object 'result' not found